fixed bugs

+++ 13/2.1.13.cpp
@@ -3,6 +3,6 @@
 int main(int argc, char** argv){
   int N, M;
   std::cin >> N >> M;
-  std::cout << ((N + M) * (N + M)) / (N * N + M * M + 1) << '\n';
+  std::cout << (double)((N + M) * (N + M)) / (N * N + M * M + 1) << '\n';
   return 0;
 }
